pyproject.toml
[project]
name = "scrapebee"
version = "0.1.0"
description = "Professional Web & Document Intelligence"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pathvalidate>=3.3.1",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "reportlab>=4.4.10",
    "selenium>=4.41.0",
    "streamlit>=1.55.0",
    "webdriver-manager>=4.0.2",
    "requests>=2.32.3",
    "python-pptx>=1.0.2",
    "pyarrow>=19.0.1",
    "fastparquet>=2024.11.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scrapebee"]
